+++ lib/vagrant/devcommands.rb
@@ -2,10 +2,7 @@ require 'vagrant'
 
 require 'vagrant/devcommands/command'
 require 'vagrant/devcommands/commandfile'
-require 'vagrant/devcommands/definer'
 require 'vagrant/devcommands/help'
 require 'vagrant/devcommands/plugin'
 require 'vagrant/devcommands/registry'
 require 'vagrant/devcommands/version'
-
-require 'vagrant/devcommands/_proxy'