+++ lib/vagrant/devcommands/_proxy.rb
@@ -1,6 +1,0 @@
-# Proxy class to make Commandfile definitions more readable.
-class VagrantDevCommands
-  def self.define(name, options)
-    VagrantPlugins::DevCommands::Definer.define(name, options)
-  end
-end